make project hooks safer, merge demo scenes interactables into sky station
